pyproject.toml
[project]
name = "paligemma-pytorch"
version = "0.1.0"
requires-python = "==3.10.*"

dependencies = [
    "fire==0.6.0",
    "gradio>=5.42.0",
    "matplotlib>=3.10.5",
    "numpy==1.26.4",
    "pillow==10.3.0",
    "safetensors==0.4.3",
    "tokenizers==0.19.1",
    "torch==2.3.0",
    "torchaudio==2.3.0",
    "torchvision==0.18.0",
    "tqdm==4.66.4",
    "transformers==4.41.2",
    "huggingface_hub>=0.24.0",
    "tensorboard>=2.20.0",
    "gradio>=5.42.0",
    "datasets>=4.0.0",
]


[tool.setuptools.packages.find]
include = ["pali_gemma*"]
